Publish image and contact code when joining community
